Init Commit
+ acft_client.c, client_phandler.py

client handler imports ctypes to help wrap my custom c functions compiled into libacft_client.dll

acft client connects to the server and sends a file.

Current Status: The handler works. The client successfully connects to and communicates with the server. The issue I'm having now is sending the .db files - Ubuntu Server isn't getting the files in a way that SQLite3 can read it as a database, I THINK it has something to do with the mass amount of garbage text being appended to the files that I send.